API Token Optimization, added instruction lining for every 60 characters, added image on the recipe information page.

+++ src/main/java/entity/Recipe.java
@@ -69,7 +69,7 @@ public class Recipe {
         this.ingredientNames = ingredientNames;
     }
 
-    public void setCalories(int calories) {
+    public void setCalories(double calories) {
         this.calories = calories;
     }
